% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00c_observations_data.R
\name{add_observations_data}
\alias{add_observations_data}
\title{Add observations data to data frame}
\usage{
add_observations_data(data, which, dataset)
}
\arguments{
\item{data}{data frame}

\item{which}{observations data}

\item{dataset}{dataset}
}
\description{
Add observations data to data frame
}
