% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables_data.R
\name{add_variables_data}
\alias{add_variables_data}
\title{Add variables data to data frame}
\usage{
add_variables_data(data, which, name, dataset)
}
\arguments{
\item{data}{data frame}

\item{which}{variables data}

\item{name}{name of new column (default = which)}

\item{dataset}{dataset}
}
\description{
Add variables data to data frame
}
