% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_handling_helpers.R
\name{alignexp_tibble_rows}
\alias{alignexp_tibble_rows}
\title{Title}
\usage{
alignexp_tibble_rows(x, by = NULL)
}
\arguments{
\item{x}{list of tibbles}

\item{by}{column to join and align}
}
\description{
Title
}
