% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_rmarkdown_functions.R
\name{button_begin}
\alias{button_begin}
\title{Creates button to hide plots and tables}
\usage{
button_begin(label = "View/Hide")
}
\arguments{
\item{label}{button label}
}
\description{
Creates button to hide plots and tables
}
