% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00e_slim_datasets.R
\name{check_data_frame}
\alias{check_data_frame}
\title{Check if the data frame name exists and if it is already loaded}
\usage{
check_data_frame(which, dataset)
}
\arguments{
\item{which}{}

\item{dataset}{}
}
\description{
Check if the data frame name exists and if it is already loaded
}
