% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00e_slim_datasets.R
\name{check_observations}
\alias{check_observations}
\title{Check if the observations has is already loaded}
\usage{
check_observations(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\description{
Check if the observations has is already loaded
}
