% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00e_slim_datasets.R
\name{check_variables}
\alias{check_variables}
\title{Check if the variables has is already loaded}
\usage{
check_variables(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\description{
Check if the variables has is already loaded
}
