% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_handling_helpers.R
\name{d2t}
\alias{d2t}
\title{Transforms any data type to a tibble}
\usage{
d2t(data, to.row.names = "rows")
}
\arguments{
\item{data}{supported data types: matrix, data.frame}

\item{to.row.names}{name for row names vector}
}
\description{
Transforms any data type to a tibble
}
