% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_databases.R
\name{db_download_STRING}
\alias{db_download_STRING}
\title{Wrapper function to download STRING databases}
\usage{
db_download_STRING(
  type = c("physical_links", "physical_links_detailed", "physical_links_full", "links",
    "links_detailed", "links_full"),
  taxid = 9606
)
}
\arguments{
\item{url}{}
}
\description{
Wrapper function to download STRING databases
}
