% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_databases.R
\name{db_download_UniProt}
\alias{db_download_UniProt}
\title{Title}
\usage{
db_download_UniProt(
  url =
    "https://rest.uniprot.org/uniprotkb/stream?fields=accession\%2Creviewed\%2Cid\%2Cprotein_name\%2Cgene_names\%2Corganism_name\%2Clength\%2Cgo\%2Cgo_p\%2Cgo_c\%2Cgo_f\%2Cxref_corum\%2Cxref_string&format=tsv&query=\%28\%28taxonomy_id\%3A10090\%29+AND+\%28reviewed\%3Atrue\%29\%29"
)
}
\arguments{
\item{url}{url decribing the anticipated columns from UniProt REST API}
}
\description{
Instruction by Henrik Hammaren
To find names of columns
go to
https://www.uniprot.org/uniprotkb?facets=reviewed%3Atrue&query=%28proteome%3AUP000005640%29
Click "Download"
Format "TSV"
Compressed "No"
Choose columns as you want
Bottom of page: "Generate URL for API"
Voila.
}
