% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_databases.R
\name{db_STRING2UniProt}
\alias{db_STRING2UniProt}
\title{Translate STRING IDs in downloaded STRING database to UniProt Accessions}
\usage{
db_STRING2UniProt(db_STRING, db_UniProt)
}
\arguments{
\item{db_STRING}{}

\item{db_UniProt}{}
}
\description{
Translate STRING IDs in downloaded STRING database to UniProt Accessions
}
