% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_databases.R
\name{db_table2sets}
\alias{db_table2sets}
\title{Prepares tables to gene set list for other functions, e.g. fgsea}
\usage{
db_table2sets(db_table, set = "Name", proteins = "UniProt")
}
\arguments{
\item{db_table}{}

\item{set}{}

\item{proteins}{}
}
\description{
Prepares tables to gene set list for other functions, e.g. fgsea
}
