% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_databases.R
\name{db_UniProt2GO}
\alias{db_UniProt2GO}
\title{Prepares a GO gene set from a downloaded UniProt table for other functions, 
e.g. fgsea}
\usage{
db_UniProt2GO(db_UniProt, ontology = c("all", "BP", "CC", "MF"))
}
\arguments{
\item{db_UniProt}{}

\item{ontology}{}
}
\description{
Prepares a GO gene set from a downloaded UniProt table for other functions, 
e.g. fgsea
}
