% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_handling_helpers.R
\name{df2t}
\alias{df2t}
\title{Transforms data frames to tibble and adds column for row names}
\usage{
df2t(data, to.row.names = "rows")
}
\arguments{
\item{data}{data frame with row names}

\item{to.row.names}{name for row names column}
}
\description{
Transforms data frames to tibble and adds column for row names
}
