% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_code_helpers.R
\name{.cat_character}
\alias{.cat_character}
\title{Prints vectors to console in as character vector}
\usage{
.cat_character(...)
}
\arguments{
\item{...}{vector/s}
}
\description{
Prints vectors to console in as character vector
}
