% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_code_helpers.R
\name{.cat_function}
\alias{.cat_function}
\title{Prints function snippet to console and clipboard}
\usage{
.cat_function(FUN, add.equal.sign = T)
}
\arguments{
\item{FUN}{function}

\item{add.equal.sign}{add equal sign to add arguments}

\item{print}{print function snippet to console}

\item{copy2clipboard}{copy function snippet to clipboard}
}
\description{
Prints function snippet to console and clipboard
}
