% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_code_helpers_get_data.R
\name{.cat_get_observations_data}
\alias{.cat_get_observations_data}
\title{Writer function for get_observations_data}
\usage{
.cat_get_observations_data(
  copy2clipboard = T,
  graphics = F,
  change_output.type = F
)
}
\arguments{
\item{copy2clipboard}{copy the code to the clipboard}

\item{graphics}{use a graphical interface for selection}

\item{change_output.type}{change output type to other than "tibble"}
}
\description{
Writer function for get_observations_data
}
