% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_code_helpers.R
\name{.cat_gsub_n}
\alias{.cat_gsub_n}
\title{Substitutes pattern multiple times and prints output as code}
\usage{
.cat_gsub_n(x, pattern, replacement, sep = "\\n\\n")
}
\arguments{
\item{x}{character string containing pattern to be substituted}

\item{pattern}{character pattern to be substituted}

\item{replacement}{character vector of strings to use for substitution}

\item{sep}{separator string between n outputs}
}
\description{
Substitutes pattern multiple times and prints output as code
}
