% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_code_helpers.R
\name{.cat_import2new_dataset}
\alias{.cat_import2new_dataset}
\title{Title}
\usage{
.cat_import2new_dataset(
  raw.data,
  variable.identifiers = "Protein.ID",
  variables.data = c("Entry.Name", "Gene", "Protein.Description", "Protein",
    "Unique.Peptides"),
  observations,
  data.frames = "channel"
)
}
\arguments{
\item{raw.data}{}

\item{variable.identifiers}{}

\item{variables.data}{}

\item{observations}{}

\item{data.frames}{}
}
\description{
Title
}
