% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_code_helpers.R
\name{.cat_mutate_from_dataframe}
\alias{.cat_mutate_from_dataframe}
\title{Define multiple mutate opereations by a formula and a corresponding data frame}
\usage{
.cat_mutate_from_dataframe(x, formula = "", copy2clipboard = T)
}
\arguments{
\item{x}{tibble with column names to be replace by each row}

\item{formula}{formula as a string containing the column names to be replaced}

\item{copy2clipboard}{copy the code to the clipboard}
}
\description{
Define multiple mutate opereations by a formula and a corresponding data frame
}
