% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_code_helpers.R
\name{.identify_observations}
\alias{.identify_observations}
\title{Guesses observations names from raw_dataset}
\usage{
.identify_observations(
  x,
  pattern = ".",
  sample.as.suffix = T,
  to.include = NULL
)
}
\arguments{
\item{x}{data frame}

\item{pattern}{separation pattern of column names}

\item{to.include}{observations that must be included; helps to identify
correct observations}
}
\description{
Guesses observations names from raw_dataset
}
