% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_structures.R
\name{.save_import}
\alias{.save_import}
\title{Saves imported files to .imports list}
\usage{
.save_import(import)
}
\arguments{
\item{import}{list of imported files}
}
\description{
Saves imported files to .imports list
}
