% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_plot_helpers.R
\name{.set_continuous_axes}
\alias{.set_continuous_axes}
\title{Defines ggplot panel by ratio, size, unit size, center, and axis breaks}
\usage{
.set_continuous_axes(
  p,
  x.axis.limits = NULL,
  y.axis.limits = NULL,
  x.axis.breaks = NULL,
  y.axis.breaks = NULL,
  aspect.ratio = 1,
  plot.center = NULL,
  axis.unit.ratio = NULL,
  expand.x.axis = c(0, 0),
  expand.y.axis = c(0, 0)
)
}
\arguments{
\item{p}{plot}

\item{x.axis.limits}{vector containing lower and upper x-axis limits}

\item{y.axis.limits}{vector containing lower and upper y-axis limits}

\item{x.axis.breaks}{distance between x-axis breaks}

\item{y.axis.breaks}{distance between y-axis breaks}

\item{aspect.ratio}{absolute length of x-axis/y-axis}

\item{plot.center}{vector for center of plot}

\item{axis.unit.ratio}{ratio between x- and y-axis units}

\item{expand.x.axis}{expand x.axis (see scale_x_continuous)}

\item{expand.y.axis}{expand x.axis (see scale_y_continuous)}
}
\description{
Defines ggplot panel by ratio, size, unit size, center, and axis breaks
}
