% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_plot_helpers.R
\name{.set_PCA_labs}
\alias{.set_PCA_labs}
\title{Title}
\usage{
.set_PCA_labs(p, PCx, PCy, sdev, digits = 1)
}
\arguments{
\item{p}{}

\item{PCx}{}

\item{PCy}{}

\item{sdev}{}

\item{digits}{}
}
\description{
Title
}
