% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_databases.R
\name{download_Reactome}
\alias{download_Reactome}
\title{Wrapper function to download data from Reactome pathways}
\usage{
download_Reactome(
  url = "https://reactome.org/download/current/UniProt2Reactome.txt",
  col_names = c("UniProt", "Reactome_Pathway_Stable_identifier", "URL", "Name",
    "Evidence_Code", "Species")
)
}
\arguments{
\item{url}{}

\item{col_names}{}
}
\description{
Wrapper function to download data from Reactome pathways
}
