% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_databases.R
\name{download_Reactome_All_Levels}
\alias{download_Reactome_All_Levels}
\title{Wrapper function to download 'All levels' data from Reactome pathways}
\usage{
download_Reactome_All_Levels(
  url = "https://reactome.org/download/current/UniProt2Reactome_All_Levels.txt",
  col_names = c("UniProt", "Reactome_Pathway_Stable_identifier", "URL", "Name",
    "Evidence_Code", "Species")
)
}
\arguments{
\item{url}{}

\item{col_names}{}
}
\description{
Wrapper function to download 'All levels' data from Reactome pathways
}
