% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_databases.R
\name{download_ReactomeReactions}
\alias{download_ReactomeReactions}
\title{Wrapper function to download 'Reactions' data from Reactome pathways}
\usage{
download_ReactomeReactions(
  url = "https://reactome.org/download/current/UniProt2ReactomeReactions.txt",
  col_names = c("UniProt", "Reactome_Pathway_Stable_identifier", "URL", "Name",
    "Evidence_Code", "Species")
)
}
\arguments{
\item{url}{}

\item{col_names}{}
}
\description{
Wrapper function to download 'Reactions' data from Reactome pathways
}
