% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_UniProt.R
\name{extract_UniProt_seqinfo}
\alias{extract_UniProt_seqinfo}
\title{Extract sequence-specific UniProt information from downloaded data}
\usage{
extract_UniProt_seqinfo(data_UniProt, info_col, remove.na = T)
}
\arguments{
\item{data_UniProt}{data frame of proteins and UniProt info as returned by 
get_UniProt_data()}

\item{info_col}{name of column to be extracted}

\item{remove.na}{remove proteins without information}
}
\description{
Extract sequence-specific UniProt information from downloaded data
}
