% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03c_fgsea.R
\name{fgsea_collapsePathways}
\alias{fgsea_collapsePathways}
\title{Title}
\usage{
fgsea_collapsePathways(
  fgseaRes,
  pathways,
  stats,
  padj.threshold = 0.01,
  pval.threshold = 0.05,
  nperm = 10/pval.threshold,
  gseaParam = 1
)
}
\arguments{
\item{fgseaRes}{Table with fgsea results}

\item{pathways}{List of pathways, should contain all the pathways present in 
'fgseaRes'.}

\item{stats}{Gene-level statistic values used for ranking, the same as in 
'fgsea()'}

\item{padj.threshold}{threshold for pathways to be considered}

\item{pval.threshold}{Two pathways are considered dependent when p-value of 
enrichment of one pathways on background of another is greater then 
  'pval.threshold'.}

\item{nperm}{Number of permutations to test for independence, should be 
several times greater than '1/pval.threhold'. Default value: 
  '10/pval.threshold'.}

\item{gseaParam}{GSEA parameter, same as for 'fgsea()'}
}
\description{
Title
}
