% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_fgsea.R
\name{fgsea_plotEnrichment}
\alias{fgsea_plotEnrichment}
\title{Title}
\usage{
fgsea_plotEnrichment(
  pathways = NULL,
  stats = NULL,
  id = "",
  title = id,
  gseaParam = 1,
  ticksSize = 0.2
)
}
\arguments{
\item{pathways}{list of gene sets used for fgsea}

\item{stats}{Gene-level stats, as in 'fgsea' function}

\item{id}{id of pathway to plot}

\item{title}{title of plot (default = <id> argument)}

\item{gseaParam}{GSEA parameter.}

\item{ticksSize}{width of vertical line corresponding to a gene (default: 0.2)}
}
\description{
Title
}
