% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03c_fgsea.R
\name{fgsea_plotGseaTable}
\alias{fgsea_plotGseaTable}
\title{Title}
\usage{
fgsea_plotGseaTable(
  n_up = 10,
  n_down = 10,
  pathways,
  stats,
  fgseaRes,
  gseaParam = 1,
  colwidths = c(5, 3, 0.8, 1.2, 1.2),
  pathwayLabelStyle = NULL,
  headerLabelStyle = NULL,
  valueStyle = NULL,
  axisLabelStyle = NULL
)
}
\arguments{
\item{n_up}{number of positively enriched pathways to show (default = 10)}

\item{n_down}{number of negatively enriched pathways to show (default = 10)}

\item{pathways}{list of gene sets used for fgsea}

\item{stats}{Gene-level stats, as in 'fgsea' function}

\item{fgseaRes}{Table with fgsea results}

\item{gseaParam}{GSEA-like parameter. Adjusts displayed statistic values, 
values closer to 0 flatten plots. Default = 1, value of 0.5 is a good choice 
too.}

\item{colwidths}{Vector of five elements corresponding to column width for 
grid.arrange. Can be both units and simple numeric vector, in latter case 
it defines proportions, not actual sizes. If column width is set to zero, 
the column is not drawn.}

\item{pathwayLabelStyle}{list with style parameter adjustments for pathway 
labels. For example, 'list(size=10, color="red")' set the font size to 10 
and color to red. See 'cowplot::draw_text' for possible options.}

\item{headerLabelStyle}{similar to 'pathwayLabelStyle' but for the table 
header.}

\item{valueStyle}{similar to 'pathwayLabelStyle' but for NES and p-value 
columns.}

\item{axisLabelStyle}{list with style parameter adjustments for stats axis 
labels. See 'ggplot2::element_text' for possible options.}
}
\description{
Title
}
