% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00d_data_frames.R
\name{get_data_frame_names}
\alias{get_data_frame_names}
\title{Prints or returns data names}
\usage{
get_data_frame_names(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\description{
Prints or returns data names
}
