% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00a_datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Checks and returns correct dataset identifier}
\usage{
get_dataset(dataset)
}
\arguments{
\item{dataset}{dataset name}
}
\description{
Checks and returns correct dataset identifier
}
