% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{Checks and returns correct dataset identifier}
\usage{
get_datasets(datasets, error = F, named = T)
}
\arguments{
\item{datasets}{dataset names}
}
\description{
Checks and returns correct dataset identifier
}
