% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00c_observations_data.R
\name{get_observations_data}
\alias{get_observations_data}
\title{Return observations data}
\usage{
get_observations_data(which, observations, output.type = "tibble", dataset)
}
\arguments{
\item{which}{which observations data to pull}

\item{observations}{(optional) vector of observations or expression}

\item{output.type}{output type (default = "tibble"; "data.frame", "vector", 
"list")}

\item{dataset}{dataset}
}
\description{
Return observations data
}
