% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00c_observations_data.R
\name{get_observations_data_names}
\alias{get_observations_data_names}
\title{Returns observations data column names}
\usage{
get_observations_data_names(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\description{
Returns observations data column names
}
