% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_databases.R
\name{get_UniProt_data}
\alias{get_UniProt_data}
\title{Download UniProt data for given protein accessions and data fields 
(see available fields with UniProt_fields())}
\usage{
get_UniProt_data(
  accession,
  fields = c("accession", "gene_names", "organism_name"),
  max.query = 1000
)
}
\arguments{
\item{accession}{vector of UniProt accessions}

\item{fields}{UniProt data fields to query}

\item{max.query}{maximum number of accessions to query at once; if the 
the number exceeds max.query, the query is split up in multiple parts}
}
\description{
Download UniProt data for given protein accessions and data fields 
(see available fields with UniProt_fields())
}
