% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables_data.R
\name{get_variables}
\alias{get_variables}
\title{Return variables}
\usage{
get_variables(variables, dataset)
}
\arguments{
\item{variables}{vector of variables}

\item{dataset}{dataset}
}
\description{
Return variables
}
