% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables_data.R
\name{get_variables_data}
\alias{get_variables_data}
\title{Return variables data}
\usage{
get_variables_data(
  which,
  variables,
  output.type = "tibble",
  output.name,
  dataset
)
}
\arguments{
\item{which}{which variables data columns to pull (multiple supported)}

\item{variables}{(optional) vector of variables or expression}

\item{output.type}{output type (default = "tibble"; "data.frame", "vector", 
"list")}

\item{dataset}{dataset}
}
\description{
Return variables data
}
