% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables_data.R
\name{get_variables_data_names}
\alias{get_variables_data_names}
\title{Returns variables data column names}
\usage{
get_variables_data_names(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\description{
Returns variables data column names
}
