% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_UniProt.R
\name{identify_UniProt_infotype}
\alias{identify_UniProt_infotype}
\title{Identify the type of UniProt data: "sequence_features" (multiple features per
protein with sequence position), "protein_feature_multiple" (multiple 
features per protein separated by "; "), or "protein_feature_single" (one 
term, feature, description per protein)}
\usage{
identify_UniProt_infotype(vector_UniProt)
}
\arguments{
\item{vector_UniProt}{vector of UniProt information like 
data_UniProt$info_type}
}
\description{
Identify the type of UniProt data: "sequence_features" (multiple features per
protein with sequence position), "protein_feature_multiple" (multiple 
features per protein separated by "; "), or "protein_feature_single" (one 
term, feature, description per protein)
}
