% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_datasets.R
\name{import2new_dataset}
\alias{import2new_dataset}
\title{Title}
\usage{
import2new_dataset(
  raw.data = NULL,
  variable.identifiers,
  variables.data = character(),
  observations = character(),
  data.frames = character()
)
}
\arguments{
\item{raw.data}{imported raw data set}

\item{variable.identifiers}{tidy expression to declare variable identifiers}

\item{variables.data}{column names from which to extract variables data}

\item{observations}{vector declaring observation names (optionally named to 
rename observations)}

\item{data.frames}{names (prefixes or suffixes) by which to identify data 
frames in the raw data sets}
}
\description{
Title
}
