% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_datasets.R
\name{import2new_dataset_long}
\alias{import2new_dataset_long}
\title{Title}
\usage{
import2new_dataset_long(
  raw.data = NULL,
  variable.identifiers,
  variables.data = character(),
  observations = character(),
  data.frames = "Precursor.Normalised",
  filter.by
)
}
\arguments{
\item{raw.data}{imported DIA-NN report file (.parquet or .tsv)}

\item{variable.identifiers}{tidy expression to declare variable identifiers}

\item{variables.data}{column names from which to extract variables data}

\item{observations}{vector declaring observation names (optionally named to 
rename observations)}

\item{data.frames}{names (prefixes or suffixes) by which to identify data 
frames in the raw data sets}

\item{proteotypic.only}{only use proteotypic precursors}

\item{Q.Value}{precursor q-value threshold}

\item{PG.Q.Value}{protein groups q-value threshold}

\item{Lib.Q.Value}{match-between-runs precursor q-value threshold}

\item{Lib.PG.Q.Value}{match-between-runs protein group q-value threshold}

\item{protein.q}{protein q-value threshold}

\item{gg.q}{gene group q-value threshold}
}
\description{
Title
}
