% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_structures.R
\name{import_files}
\alias{import_files}
\title{Imports any file type using the file extension and returns list}
\usage{
import_files(files, silent = F)
}
\arguments{
\item{files}{file paths}

\item{silent}{suppresses messages}

\item{dir}{directory to import files from}

\item{ext}{specific file extensions to imports}
}
\description{
Imports any file type using the file extension and returns list
}
