% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Data_transformation.R
\name{limma_normalizeBetweenArrays}
\alias{limma_normalizeBetweenArrays}
\title{limma's normalizeBetweenArrays method for regular tibble with variables as 
columns}
\usage{
limma_normalizeBetweenArrays(
  object,
  method = "scale",
  targets = NULL,
  cyclic.method = "fast",
  ...
)
}
\arguments{
\item{object}{}

\item{method}{}

\item{targets}{}

\item{cyclic.method}{}

\item{...}{}
}
\description{
limma's normalizeBetweenArrays method for regular tibble with variables as 
columns
}
