% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_handling_helpers.R
\name{list2logical_df}
\alias{list2logical_df}
\title{Transforms a list to tibble logical indication for variables}
\usage{
list2logical_df(x, identifier = "variables")
}
\arguments{
\item{x}{list}
}
\description{
Transforms a list to tibble logical indication for variables
}
