% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_handling_helpers.R
\name{matrix2tibble}
\alias{matrix2tibble}
\title{Transforms matrix to tibble and adds row names as first column}
\usage{
matrix2tibble(data, to.row.names = "rows")
}
\arguments{
\item{data}{matrix with row names}

\item{to.row.names}{name for row names column}
}
\description{
Transforms matrix to tibble and adds row names as first column
}
