% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_fasta.R
\name{mutate_sequence_position}
\alias{mutate_sequence_position}
\title{Title}
\usage{
mutate_sequence_position(
  x,
  fasta,
  identifier = "Protein.Group",
  sequence = "Stripped.Sequence"
)
}
\arguments{
\item{x}{tibble}

\item{fasta}{location of a fasta file}

\item{identifier}{column name of protein/groups id}

\item{sequence}{column name of peptide sequences}
}
\description{
Title
}
