% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_datasets.R
\name{new_dataset}
\alias{new_dataset}
\title{Title}
\usage{
new_dataset(variables_data_frame, observations_data, data_frame_list)
}
\arguments{
\item{variables_data_frame}{tibble with "variables" as first column and 
additional after}

\item{observations_data}{tibble with "observations" as first column and 
additional after}

\item{data_frame_list}{named list of data frames}
}
\description{
Title
}
