% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_structures.R
\name{NGLVieweR_AFdb}
\alias{NGLVieweR_AFdb}
\title{NGLVieweR function which accepts protein accession IDs and downloads given 
protein structure from the Alphafold database}
\usage{
NGLVieweR_AFdb(id)
}
\arguments{
\item{id}{}
}
\description{
NGLVieweR function which accepts protein accession IDs and downloads given 
protein structure from the Alphafold database
}
