% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_rmarkdown_functions.R
\name{nice_number}
\alias{nice_number}
\title{Help make nice number in R Markdown}
\usage{
nice_number(
  x,
  round = T,
  digits = 0,
  format = "d",
  big.mark = ",",
  sci.digits,
  output = "numeric"
)
}
\arguments{
\item{x}{number}

\item{round}{should number be rounded}

\item{digits}{round to number of digits after comma}

\item{format}{number format (see formatC())}

\item{big.mark}{separator between 3 digits; remove with ""}

\item{sci.digits}{digits after decimal point for scientific notation
(format = "e)}

\item{output}{type of output vector}
}
\description{
Help make nice number in R Markdown
}
