% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Data_transformation.R
\name{normalizeBetweenArrays_subset}
\alias{normalizeBetweenArrays_subset}
\title{Applies limma::normalizeBetweenArrays to an matrix but calculates 
normalization factors only based on a subset of the proteins}
\usage{
normalizeBetweenArrays_subset(
  object,
  method,
  targets,
  cyclic.method,
  reference_proteins,
  ...
)
}
\arguments{
\item{object}{}

\item{method}{}

\item{targets}{}

\item{cyclic.method}{}

\item{reference_proteins}{}

\item{...}{}
}
\description{
Applies limma::normalizeBetweenArrays to an matrix but calculates 
normalization factors only based on a subset of the proteins
}
