% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_handling_helpers.R
\name{objects_to_list}
\alias{objects_to_list}
\title{Add objects by name to a list and (optionally) delete them}
\usage{
objects_to_list(object.names, x = list(), rm.objects = T)
}
\arguments{
\item{object.names}{names of objects to add to list}

\item{x}{list to assign objects to}

\item{rm.objects}{}
}
\description{
Add objects by name to a list and (optionally) delete them
}
