% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00d_data_frames.R
\name{save_data_frame}
\alias{save_data_frame}
\title{Adds new data}
\usage{
save_data_frame(data_frame, name, dataset)
}
\arguments{
\item{data_frame}{new data}

\item{name}{data name}

\item{dataset}{dataset name
name}
}
\description{
Adds new data
}
