% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00c_observations_data.R
\name{save_observations_data}
\alias{save_observations_data}
\title{Adds data to observations data frame}
\usage{
save_observations_data(data_frame, column, name, dataset, fill = NA)
}
\arguments{
\item{data_frame}{observations data as tibble or named vector}

\item{column}{column names to save}

\item{name}{name}

\item{dataset}{dataset}

\item{fill}{value to fill non-existing observations with}
}
\description{
Adds data to observations data frame
}
