% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables_data.R
\name{save_variables_data}
\alias{save_variables_data}
\title{Add data to variables data frame}
\usage{
save_variables_data(data_frame, column, dataset, fill = NA)
}
\arguments{
\item{data_frame}{variables data as tibble or named vector}

\item{column}{column names to save}

\item{dataset}{dataset}

\item{fill}{value to fill non-existing variables with}
}
\description{
Add data to variables data frame
}
