% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_strsplit_helpers.R
\name{str_locate_last}
\alias{str_locate_last}
\title{Locates last position of pattern in given string}
\usage{
str_locate_last(string, pattern)
}
\arguments{
\item{string}{vector of strings}

\item{pattern}{pattern to match}
}
\description{
Locates last position of pattern in given string
}
