% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_strsplit_helpers.R
\name{strsplit_}
\alias{strsplit_}
\title{Same as strsplit but option to return as vector and to convert to numeric}
\usage{
strsplit_(x, split, vector.output = T, as_numeric = F)
}
\arguments{
\item{x}{string}

\item{split}{split character}

\item{vector.output}{output type (T = vector; F = list)}

\item{as.numeric}{try to convert to numeric}
}
\description{
Same as strsplit but option to return as vector and to convert to numeric
}
