% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_UniProt.R
\name{T2G_2list}
\alias{T2G_2list}
\title{Transforms a TERM2GENE data frame into a list of Term sets}
\usage{
T2G_2list(data_T2G, info_col = 1, entry_col = "Entry")
}
\arguments{
\item{data_T2G}{TERM2GENE data frame}

\item{info_col}{Term column}

\item{entry_col}{Gene column}
}
\description{
Transforms a TERM2GENE data frame into a list of Term sets
}
