% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_handling_helpers.R
\name{t2m}
\alias{t2m}
\title{Transforms tibble to matrix}
\usage{
t2m(data)

t2m(data)
}
\arguments{
\item{data}{tibble}
}
\description{
Transforms tibble to matrix

Transforms tibble to matrix
}
