% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_rmarkdown_functions.R
\name{template2Rmd_fill}
\alias{template2Rmd_fill}
\title{Specific function for 00_Master script to fill Rmd template and add content}
\usage{
template2Rmd_fill(
  name,
  dir = "Scripts",
  dir.out,
  template = "00_template.Rmd",
  fill.pattern = "# Content",
  output.name
)
}
\arguments{
\item{name}{name of script and that should be replaced in the template file}

\item{dir}{directory of the template}

\item{dir.out}{output directory of the Rmd file}

\item{template}{name of the Rmd template}

\item{fill.pattern}{(optional) pattern to be replaced with existing Rmd 
script specified by name}

\item{output.name}{name of output Rmd file (defaults to gsub("00_template", 
name, template))}
}
\description{
Specific function for 00_Master script to fill Rmd template and add content
}
