% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00e_slim_datasets.R
\name{thicc_datasets}
\alias{thicc_datasets}
\title{Load all data frames, variables data, and observaitons data into Datasets 
list}
\usage{
thicc_datasets()
}
\description{
Load all data frames, variables data, and observaitons data into Datasets 
list
}
