% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00e_slim_datasets.R
\name{thicc_observations_data}
\alias{thicc_observations_data}
\title{Load all observaitons data into Datasets list}
\usage{
thicc_observations_data()
}
\description{
Load all observaitons data into Datasets list
}
