% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_handling_helpers.R
\name{tibble2data_frame}
\alias{tibble2data_frame}
\title{Transforms tibble to data frame}
\usage{
tibble2data_frame(data)
}
\arguments{
\item{data}{tibble}
}
\description{
Transforms tibble to data frame
}
