% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_code_helpers.R
\name{cleanup}
\alias{cleanup}
\title{Remove objects from the global environment and the Imports list}
\usage{
cleanup(global = T, exclude = c(), imports = T)
}
\arguments{
\item{global}{should the global environment be cleaned}

\item{exclude}{which objects not to remove}

\item{imports}{should the imported data frames be removed}
}
\description{
Remove objects from the global environment and the Imports list
}
