% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00d_data_frames.R
\name{get_data_frame_m}
\alias{get_data_frame_m}
\title{Assemble data frames from dataset and return as list}
\usage{
get_data_frame_m(
  which,
  variables,
  observations,
  dataset,
  output.type = "tibble"
)
}
\arguments{
\item{which}{specific names of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{dataset}{dataset name}

\item{output.type}{data type (default = "tibble", "data.frame", "matrix")}
}
\description{
Assemble data frames from dataset and return as list
}
