% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables_data.R
\name{get_variables_data}
\alias{get_variables_data}
\title{Return variables data}
\usage{
get_variables_data(which, variables, dataset, output.type = "tibble")
}
\arguments{
\item{which}{which variables data columns to pull (multiple supported)}

\item{variables}{(optional) vector of variables or expression}

\item{dataset}{dataset}

\item{output.type}{output type (default = "tibble"; "data.frame", "vector", 
"list")}
}
\description{
Return variables data
}
